Refactor: 변경 영화 리스트 컴포넌트 위치
공통 컴포넌트이기 때문에 컴포넌트 폴더로 이동
