+++ src/components/MovieList/MovieList.jsx
@@ -1,0 +1,16 @@
+import MovieItem from './MovieItem';
+import { Container, MovieImageList } from './MovieList.style';
+
+const MovieList = ({ movies }) => {
+  return (
+    <Container>
+      <MovieImageList gap={12}>
+        {movies?.map(({ id, title, vote_average, poster_path }) => (
+          <MovieItem key={id} title={title} vote_average={vote_average} poster_path={poster_path} />
+        ))}
+      </MovieImageList>
+    </Container>
+  );
+};
+
+export default MovieList;
