+++ src/components/Star/Star.jsx
