+++ src/components/Star/Star.style.js
