+++ src/pages/Search/Search.jsx
@@ -3,7 +3,7 @@ import { useQuery } from 'react-query';
 
 import { getSearchMovie } from '../../api/api.js';
 import { Bold, Box, Title } from './Search.style.js';
-import MovieList from './MovieList/MovieList.jsx';
+import MovieList from '../../components/MovieList/MovieList.jsx';
 import { Backdrop, CircularProgress } from '@mui/material';
 
 const Search = () => {
@@ -12,7 +12,6 @@ const Search = () => {
   const { data: movies, status } = useQuery(['searchMovie', query], () => getSearchMovie(query));
 
   if (status === 'loading') {
-    // return <>loading...</>;
     return (
       <Backdrop sx={{ color: '#fff', zIndex: theme => theme.zIndex.drawer + 1 }} open={true}>
         <CircularProgress color="inherit" />
